+++ packages/graphql-language-service-server/src/GraphQLCache.ts
@@ -54,16 +54,18 @@ const {
 
 export async function getGraphQLCache(
   configDir: Uri,
+  parser: typeof parseDocument,
   extensions?: Array<(config: GraphQLConfig) => GraphQLConfig>,
   config?: GraphQLConfig,
 ): Promise<GraphQLCacheInterface> {
-  let graphQLConfig = config || (await loadConfig({ rootDir: configDir }));
+  let graphQLConfig =
+    config ?? ((await loadConfig({ rootDir: configDir })) as GraphQLConfig);
   if (extensions && extensions.length > 0) {
     for await (const extension of extensions) {
       graphQLConfig = await extension(graphQLConfig);
     }
   }
-  return new GraphQLCache(configDir, graphQLConfig);
+  return new GraphQLCache(configDir, graphQLConfig, parser);
 }
 
 export class GraphQLCache implements GraphQLCacheInterface {
@@ -74,8 +76,13 @@ export class GraphQLCache implements GraphQLCacheInterface {
   _typeExtensionMap: Map<Uri, number>;
   _fragmentDefinitionsCache: Map<Uri, Map<string, FragmentInfo>>;
   _typeDefinitionsCache: Map<Uri, Map<string, ObjectTypeInfo>>;
+  _parser: typeof parseDocument;
 
-  constructor(configDir: Uri, graphQLConfig: GraphQLConfig) {
+  constructor(
+    configDir: Uri,
+    graphQLConfig: GraphQLConfig,
+    parser: typeof parseDocument,
+  ) {
     this._configDir = configDir;
     this._graphQLConfig = graphQLConfig;
     this._graphQLFileListCache = new Map();
@@ -83,6 +90,7 @@ export class GraphQLCache implements GraphQLCacheInterface {
     this._fragmentDefinitionsCache = new Map();
     this._typeDefinitionsCache = new Map();
     this._typeExtensionMap = new Map();
+    this._parser = parser;
   }
 
   getGraphQLConfig = (): GraphQLConfig => this._graphQLConfig;
@@ -796,7 +804,7 @@ export class GraphQLCache implements GraphQLCacheInterface {
         let queries: CachedContent[] = [];
         if (content.trim().length !== 0) {
           try {
-            queries = parseDocument(content, filePath);
+            queries = this._parser(content, filePath);
             if (queries.length === 0) {
               // still resolve with an empty ast
               resolve({